requirements.txt
aiohttp==3.8.5
PyYAML==6.0
requests==2.31.0
